Organize assets into data and image folders
