+++ assets/data/elemental_relations.ts
