+++ assets/data/spells.js
