+++ tsconfig.json
@@ -7,5 +7,5 @@
     "strict": false,
     "esModuleInterop": true
   },
-  "include": ["*.ts"]
+  "include": ["*.ts", "assets/data/**/*.ts"]
 }
